pyproject.toml
[tool.poetry]
name = "clfuncspoc"
version = "0.0.0"
description = "Chainlink Functions Proof of Concept"
authors = ["Gosuto Inzasheru <gosuto@inzasheru.email>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.10"
eth-brownie = "^1.19.3"
rich = "^13.4.2"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
